Add Piggyvault to Open Source Projects
This is one of my side projects which basically enables users to log their daily transactions and later this data can be used to get insights. It has all the major modules required for family finance management like accounts, categories, multi-currency support, etc. It is a multi-tenant app, and each tenant is considered as a family. 

The project also includes a backend project built on .NET Core and SQL Server is used as the database.